fix(core): fix TypeScript typings
Fixes #11
